+++ include/ZBE/archetypes/implementations/SimpleWideMobile.h
@@ -10,8 +10,6 @@
 #ifndef ZBE_ARCHETYPES_IMPLEMENTATIONS_SIMPLEWIDEMOBILE_H
 #define ZBE_ARCHETYPES_IMPLEENTATIONS_SIMPLEWIDEMOBILE_H
 
-#include <cstdint>
-
 #include "ZBE/archetypes/WideMobile.h"
 #include "ZBE/archetypes/implementations/SimpleMobile.h"
 
@@ -20,20 +18,20 @@ namespace zbe {
 template <unsigned s>
 class SimpleWideMobile : virtual public WideMobile<s>, public SimpleMobile<s> {
   public:
-    SimpleWideMobile(uint64_t width) : SimpleMobile<s>(), w(width) {}
-    SimpleWideMobile(std::initializer_list<int64_t> position, uint64_t width) : SimpleMobile<s>(position), w(width) {}
-    SimpleWideMobile(Point<s> position, uint64_t width) : SimpleMobile<s>(position), w(width) {}
-    SimpleWideMobile(Vector<s> velocity, uint64_t width) : SimpleMobile<s>(velocity), w(width) {}
-    SimpleWideMobile(std::initializer_list<int64_t> position, std::initializer_list<int64_t> velocity, uint64_t width) : SimpleMobile<s>(position, velocity), w(width) {}
-    SimpleWideMobile(Point<s> position, std::initializer_list<int64_t> velocity, uint64_t width) : SimpleMobile<s>(position, velocity), w(width) {}
-    SimpleWideMobile(Point<s> position, Vector<s> velocity, uint64_t width) : SimpleMobile<s>(position, velocity), w(width) {}
-    SimpleWideMobile(std::initializer_list<int64_t> position, Vector<s> velocity, uint64_t width) : SimpleMobile<s>(position, velocity), w(width) {}
+    SimpleWideMobile(double width) : SimpleMobile<s>(), w(width) {}
+    SimpleWideMobile(std::initializer_list<double> position, double width) : SimpleMobile<s>(position), w(width) {}
+    SimpleWideMobile(Point<s> position, double width) : SimpleMobile<s>(position), w(width) {}
+    SimpleWideMobile(Vector<s> velocity, double width) : SimpleMobile<s>(velocity), w(width) {}
+    SimpleWideMobile(std::initializer_list<double> position, std::initializer_list<double> velocity, double width) : SimpleMobile<s>(position, velocity), w(width) {}
+    SimpleWideMobile(Point<s> position, std::initializer_list<double> velocity, double width) : SimpleMobile<s>(position, velocity), w(width) {}
+    SimpleWideMobile(Point<s> position, Vector<s> velocity, double width) : SimpleMobile<s>(position, velocity), w(width) {}
+    SimpleWideMobile(std::initializer_list<double> position, Vector<s> velocity, double width) : SimpleMobile<s>(position, velocity), w(width) {}
 
-    void setWidth(uint64_t width) {w = width;}
-    uint64_t getWidth() {return (w);}
+    void setWidth(double width) {w = width;}
+    double getWidth() {return (w);}
 
   private:
-    uint64_t w;
+    double w;
 };
 
 }  // namespace
