+++ include/ZBE/entities/avatars/Positionable.h
@@ -20,7 +20,7 @@ class Positionable {
   public:
     virtual ~Positionable() {}
 
-    virtual void setPosition(std::initializer_list<int64_t> l) = 0;
+    virtual void setPosition(std::initializer_list<double> l) = 0;
     virtual void setPosition(Point<s> position) = 0;
 
     virtual Point<s>& getPosition() = 0;
